requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
pydantic
